fixed bugs and added signals to delete media files

+++ MetioTube/core/auto_delete_account_if_not_activated.py
@@ -7,8 +7,10 @@ UserModel = get_user_model()
 
 def auto_delete_account_if_not_activated(user_id):
     time.sleep(300)
+    user = UserModel.objects.filter(pk=user_id).get()
 
-    user = UserModel.objects.get(pk=user_id)
+    if not user:
+        return
 
     if not user.is_active:
         user.delete()
